+++ NEWS.md
@@ -7,6 +7,7 @@
   - `position_group` - Position group of the player
   - `headshot_url` - URL to a player headshot image
 * Fix bad possession team info in BUF-LA
+* `decode_player_ids()` uses `nflreadr::load_players()` to translate player IDs to gsis IDs starting in the 2022 season.
 
 # nflfastR 4.4.0